rag/pyproject.toml
[tool.poetry]
name = "rag"
version = "0.1.0"
description = ""
authors = ["WJJblack1224 "]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.3.0"
python-dotenv = "^1.0.1"
langchain-openai = "^0.2.0"
langchain-community = "^0.3.0"
pypdf = "^4.3.1"
langchain-qdrant = "0.2.0.dev1"
google-generativeai = "^0.7.0"
langchain-google-genai = "^2.0.0"
ipython = "^8.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
